Move ValueInSingleWindow from testing package to values package
