+++ sdks/java/core/src/main/java/org/apache/beam/sdk/testing/GatherAllPanes.java
@@ -31,6 +31,7 @@ import org.apache.beam.sdk.util.IdentityWindowFn;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.TypeDescriptor;
+import org.apache.beam.sdk.values.ValueInSingleWindow;
 
 /**
  * Gathers all panes of each window into exactly one output.
