+++ sdks/java/core/src/main/java/org/apache/beam/sdk/testing/PaneExtractors.java
@@ -27,6 +27,7 @@ import org.apache.beam.sdk.transforms.windowing.PaneInfo;
 import org.apache.beam.sdk.transforms.windowing.PaneInfo.Timing;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.TypeDescriptor;
+import org.apache.beam.sdk.values.ValueInSingleWindow;
 
 /**
  * {@link PTransform PTransforms} which take an {@link Iterable} of {@link ValueInSingleWindow
